Fixed 2 bugs in the admin reports: 1. percent complete on the progress report was not being calculated and displayed correctly 2. categories with a false value were displaying blank instead of false on the invitation report The 4 reports are all displaying correct data now

+++ Services/CategoryInfo.cs
@@ -10,7 +10,7 @@ namespace SinStim.Services {
             this.Category = category;
             this.TotalPictures = totalPictures;
             this.FinishedPictureCount = finishedPictureCount;
-            this.PercentComplete = this.FinishedPictureCount / this.TotalPictures;
+            this.PercentComplete = (double)this.FinishedPictureCount / (double)this.TotalPictures;
         }
     }
 }
